+++ src/at.d.ts
@@ -1,0 +1,6 @@
+import type {ImmutableTree} from './types';
+
+export default function at<T>(
+  tree: ImmutableTree<T>,
+  index: number,
+): T;