+++ src/errors.mjs
@@ -1,5 +1,14 @@
 // @flow strict
 
+export class IndexOutOfRangeError extends Error {
+  constructor(value/*: mixed */) {
+    super(
+      'The given tree is out of range: ' +
+      String(value),
+    );
+  }
+}
+
 export class ValueExistsError extends Error {
   constructor(value/*: mixed */) {
     super(