+++ src/index.d.ts
@@ -1,3 +1,4 @@
+export {default as at} from './at';
 export {default as create} from './create';
 export {default as difference} from './difference';
 export {default as equals} from './equals';