+++ src/index.mjs
@@ -1,5 +1,6 @@
 // @flow strict
 
+export {default as at} from './at.mjs';
 export {default as create} from './create.mjs';
 export {default as difference} from './difference.mjs';
 export {default as equals} from './equals.mjs';